refactor!: ring_buffers now shared_ring_buffers
naming and namespace change for backing ring buffer, now prefix with
"shared", this extends to the shared_ring_buffer_context header and all
functions contained (going from "rb" to "srb"). This is to make space
for a new generic ring buffer header in the same style as the vector
header.